requirements.txt
streamlit
pandas
numpy
plotly.express
datetime
scipy.cluster.hierarchy
matplotlib
wordcloud